norm

+++ philo_three/utils/ft_gettime.c
@@ -1,3 +1,15 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   ft_gettime.c                                       :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: seolim <seolim@student.42.fr>              +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2021/02/01 16:35:12 by seolim            #+#    #+#             */
+/*   Updated: 2021/02/01 16:40:29 by seolim           ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
 #include "utils.h"
 
 long		ft_gettime(void)
@@ -5,5 +17,5 @@ long		ft_gettime(void)
 	static struct timeval	tv;
 
 	gettimeofday(&tv, NULL);
-	return ((int)((tv.tv_sec * 1000) + (tv.tv_usec / 1000)));
+	return ((tv.tv_sec * 1000) + (tv.tv_usec / 1000));
 }
